refactor logging and password providers into their own module
