+++ src/main/python/afp_cli/log.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+from __future__ import print_function, absolute_import, division
+
+
+import sys
+
+
+DEBUG = False
+
+
+def error(message):
+    print(message, file=sys.stderr)
+    sys.exit(1)
+
+
+def debug(message):
+    if DEBUG:
+        print(message)
